# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ice Truck Tracking API",
    "version": "2.0.0",
    "description": "Professional Ice Truck Tracking System API — PostgreSQL + TimescaleDB, Redis, MQTT"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Driver": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "license_number": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Truck": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "plate_number": { "type": "string" },
          "model": { "type": "string" },
          "capacity": { "type": "number" },
          "status": { "type": "string", "enum": ["active", "inactive", "maintenance"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Shop": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "address": { "type": "string" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Alert": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "type": { "type": "string" },
          "message": { "type": "string" },
          "severity": { "type": "string", "enum": ["low", "medium", "high", "critical"] },
          "truck_id": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/drivers": {
      "get": {
        "tags": ["Drivers"],
        "summary": "Get all drivers",
        "responses": {
          "200": {
            "description": "List of drivers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "results": { "type": "integer" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Driver" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Drivers"],
        "summary": "Create a driver (requires drivers:create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Driver" }
            }
          }
        },
        "responses": {
          "201": { "description": "Driver created" },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
    "/drivers/{id}": {
      "get": {
        "tags": ["Drivers"],
        "summary": "Get driver by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Driver details" },
          "404": { "description": "Driver not found" }
        }
      }
    },
    "/trucks": {
      "get": {
        "tags": ["Trucks"],
        "summary": "Get all trucks",
        "responses": {
          "200": {
            "description": "List of trucks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "results": { "type": "integer" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Truck" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Trucks"],
        "summary": "Create a truck (requires trucks:create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Truck" }
            }
          }
        },
        "responses": {
          "201": { "description": "Truck created" },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
    "/trucks/{id}": {
      "get": {
        "tags": ["Trucks"],
        "summary": "Get truck by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Truck details" },
          "404": { "description": "Truck not found" }
        }
      }
    },
    "/shops": {
      "get": {
        "tags": ["Shops"],
        "summary": "Get all shops",
        "responses": {
          "200": {
            "description": "List of shops",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "results": { "type": "integer" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Shop" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Shops"],
        "summary": "Create a shop (requires shops:create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Shop" }
            }
          }
        },
        "responses": {
          "201": { "description": "Shop created" },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
    "/shops/{id}": {
      "get": {
        "tags": ["Shops"],
        "summary": "Get shop by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Shop details" },
          "404": { "description": "Shop not found" }
        }
      }
    },
    "/alerts": {
      "get": {
        "tags": ["Alerts"],
        "summary": "Get all alerts",
        "responses": {
          "200": {
            "description": "List of alerts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "results": { "type": "integer" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Alert" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Alerts"],
        "summary": "Create an alert (requires alerts:create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Alert" }
            }
          }
        },
        "responses": {
          "201": { "description": "Alert created" },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
...
  }
}